cbfs: 64-bit cleanups
cbfs used u32 in a number of cases where uintptr_t was
correct. This change builds for both 64-bit and 32-bit
boards.

Also, add a function to convert 64-bit struct members
to a void *. Note that it accomodates the incorrect
byte ordering from cbfs -- which has been prevalent
in coreboot since the beginning, and which we will not
fix this time around. People might be upset if they
suddenly can't read their flash images with cbfs.

+++ src/lib/cbmem.c
@@ -200,7 +200,7 @@ void *cbmem_add(u32 id, u64 size)
 	cbmem_toc[0].base += size;
 	cbmem_toc[0].size -= size;
 
-	return (void *)(u32)cbmem_toc[i].base;
+	return (void *)(uintptr_t)cbmem_toc[i].base;
 }
 
 void *cbmem_find(u32 id)
